+++ packages/ai/src/coder/index.ts
@@ -1,1 +1,1 @@
-export * from './coder';
+export * from './block';
