+++ packages/ai/test/coder/diffs/data/multiple/before.txt
