+++ packages/ai/test/coder/diffs/data/multiple/diff.txt
@@ -1,0 +1,25 @@
+file1.ts
+```typescript
+<<<<<<< SEARCH
+      1
+=======
+      One
+>>>>>>> REPLACE
+```
+
+```javascript
+<<<<<<< SEARCH
+      2
+=======
+      Two
+>>>>>>> REPLACE
+```
+
+src/file2.tsx
+```
+<<<<<<< SEARCH
+      3
+=======
+      Three
+>>>>>>> REPLACE
+```
