+++ packages/ai/test/coder/diffs/data/single/after.txt
@@ -1,0 +1,3 @@
+<>
+      BAR
+</>
