+++ packages/ai/test/coder/diffs/data/single/before.txt
@@ -1,0 +1,3 @@
+<>
+      FOO
+</>
