+++ packages/ai/test/coder/diffs/data/single/diff.txt
@@ -1,5 +1,7 @@
+123
 <<<<<<< SEARCH
       FOO
 =======
       BAR
 >>>>>>> REPLACE
+xyz
