+++ packages/models/src/chat/message/code.ts
@@ -1,0 +1,5 @@
+export interface CodeBlock {
+    fileName?: string;
+    language?: string;
+    content: string;
+}
