+++ packages/models/src/chat/message/index.ts
@@ -1,3 +1,4 @@
 export * from '../response';
+export * from './code';
 export * from './context';
 export * from './message';
